🐛 fix: 상승/하락 순 쿼리를 하나의 쿼리로 묶어서 호출(분기처리)
